It's my BCA 6th semester project.
This project on vb.net based website which facilated for gives the needs
of daily life like teacher problem of your children,  event handling ,
ambulance on emergency , good shipment , technician for car, electricity
and computer, house keeper and mobile recharge.

Kalpataru is a wish-fulfilling divine tree in Hindu mythology, also
known as Kalpavriksha.

The name Kalpataru is applied because this site also like the
Kalpavriksha , that fulfil your daily wishes .